Corrected name of auth_service package
