+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_service/config/AuthConfig.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.user_auth_service.auth_sevice.config;
+package com.crmfoodestablishment.user_auth_service.auth_service.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
