+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_service/service/impl/AuthServiceImpl.java
@@ -1,15 +1,15 @@
-package com.crmfoodestablishment.user_auth_service.auth_sevice.service.impl;
+package com.crmfoodestablishment.user_auth_service.auth_service.service.impl;
 
-import com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload.LoginRequestPayload;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload.RegisterResponsePayload;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload.TokenPairResponsePayload;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload.UserRegistrationRequestPayload;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.exception.FailedRegistrationException;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.exception.InvalidTokenException;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.exception.InvalidUserCredentialsException;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.service.AuthService;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.service.JwtService;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.service.token.RefreshToken;
+import com.crmfoodestablishment.user_auth_service.auth_service.controller.payload.LoginRequestPayload;
+import com.crmfoodestablishment.user_auth_service.auth_service.controller.payload.RegisterResponsePayload;
+import com.crmfoodestablishment.user_auth_service.auth_service.controller.payload.TokenPairResponsePayload;
+import com.crmfoodestablishment.user_auth_service.auth_service.controller.payload.UserRegistrationRequestPayload;
+import com.crmfoodestablishment.user_auth_service.auth_service.exception.FailedRegistrationException;
+import com.crmfoodestablishment.user_auth_service.auth_service.exception.InvalidTokenException;
+import com.crmfoodestablishment.user_auth_service.auth_service.exception.InvalidUserCredentialsException;
+import com.crmfoodestablishment.user_auth_service.auth_service.service.AuthService;
+import com.crmfoodestablishment.user_auth_service.auth_service.service.JwtService;
+import com.crmfoodestablishment.user_auth_service.auth_service.service.token.RefreshToken;
 
 import com.crmfoodestablishment.user_auth_service.user_manager.controller.UserController;
 import com.crmfoodestablishment.user_auth_service.user_manager.exception.NotFoundException;
