+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_service/service/token/AccessToken.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.user_auth_service.auth_sevice.service.token;
+package com.crmfoodestablishment.user_auth_service.auth_service.service.token;
 
 import io.jsonwebtoken.Header;
