+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_service/service/token/RefreshTokenClaims.java
@@ -1,4 +1,4 @@
-package com.crmfoodestablishment.user_auth_service.auth_sevice.service.token;
+package com.crmfoodestablishment.user_auth_service.auth_service.service.token;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
