+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_service/service/token/adapter/AccessTokenHandlerAdapter.java
@@ -1,7 +1,7 @@
-package com.crmfoodestablishment.user_auth_service.auth_sevice.service.token.adapter;
+package com.crmfoodestablishment.user_auth_service.auth_service.service.token.adapter;
 
-import com.crmfoodestablishment.user_auth_service.auth_sevice.service.token.AccessToken;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.service.token.AccessTokenClaims;
+import com.crmfoodestablishment.user_auth_service.auth_service.service.token.AccessToken;
+import com.crmfoodestablishment.user_auth_service.auth_service.service.token.AccessTokenClaims;
 import com.crmfoodestablishment.user_auth_service.user_manager.entity.Permission;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
@@ -10,7 +10,7 @@ import io.jsonwebtoken.JwtHandlerAdapter;
 import java.util.ArrayList;
 import java.util.UUID;
 
-import static com.crmfoodestablishment.user_auth_service.auth_sevice.service.TimeUtils.convertDateToLocalDateTime;
+import static com.crmfoodestablishment.user_auth_service.auth_service.service.TimeUtils.convertDateToLocalDateTime;
 
 public class AccessTokenHandlerAdapter extends JwtHandlerAdapter<AccessToken> {
