+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/auth_service/service/token/adapter/RefreshTokenHandlerAdapter.java
@@ -1,14 +1,14 @@
-package com.crmfoodestablishment.user_auth_service.auth_sevice.service.token.adapter;
+package com.crmfoodestablishment.user_auth_service.auth_service.service.token.adapter;
 
-import com.crmfoodestablishment.user_auth_service.auth_sevice.service.token.RefreshToken;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.service.token.RefreshTokenClaims;
+import com.crmfoodestablishment.user_auth_service.auth_service.service.token.RefreshToken;
+import com.crmfoodestablishment.user_auth_service.auth_service.service.token.RefreshTokenClaims;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jws;
 import io.jsonwebtoken.JwtHandlerAdapter;
 
 import java.util.UUID;
 
-import static com.crmfoodestablishment.user_auth_service.auth_sevice.service.TimeUtils.convertDateToLocalDateTime;
+import static com.crmfoodestablishment.user_auth_service.auth_service.service.TimeUtils.convertDateToLocalDateTime;
 
 public class RefreshTokenHandlerAdapter extends JwtHandlerAdapter<RefreshToken> {
