+++ user-manager-service/src/main/java/com/crmfoodestablishment/user_auth_service/user_manager/services/UserService.java
@@ -1,7 +1,7 @@
 package com.crmfoodestablishment.user_auth_service.user_manager.services;
 
 import com.crmfoodestablishment.user_auth_service.user_manager.entity.User;
-import com.crmfoodestablishment.user_auth_service.auth_sevice.controller.payload.UserRegistrationRequestPayload;
+import com.crmfoodestablishment.user_auth_service.auth_service.controller.payload.UserRegistrationRequestPayload;
 
 import java.util.UUID;
